refactor: adopt compound components pattern
