+++ packages/components/workbench/src/index.ts
@@ -1,2 +1,2 @@
-export { Workbench } from './Workbench';
+export { Workbench } from './CompoundWorkbench';
 export type { WorkbenchProps } from './Workbench';
